requirements.txt
azure-log-analytics-data-collector-api
requests
flask